[UPD] splitted up scheduler into 3 much smaller files for readability, core, actions and reactions
