+++ server/scheduler/__init__.py
@@ -1,0 +1,3 @@
+from .core import init_scheduler, shutdown_scheduler
+
+__all__ = ['init_scheduler', 'shutdown_scheduler']
